+++ test/faker/color_test.exs
@@ -1,0 +1,9 @@
+defmodule Faker.ColorTest do
+  use ExUnit.Case, async: true
+  import Faker.Color
+
+  test :rgb_hex do
+    assert String.length(rgb_hex) == 6
+    assert Regex.match?(~r/^[0-9A-F]{6}$/, rgb_hex)
+  end
+end